Reverted some of the code
